pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promlint"
version = "0.1.0"
description = "Lint Prometheus alerting rule files for common quality issues and anti-patterns."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Matt Gacek" }]
dependencies = [
    "ruamel.yaml>=0.18",
    "promql-parser>=0.6",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
promlint = "promlint.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/promlint"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
